refactor(linter): rename `flat.rs` to `config.rs` (#8033)
simply renames `flat.rs` to `config.rs` to more accurately reflect the file's purpose
